this commit's YUGE, folks!
Restructured EVERYTHING! If it had code, it was changed. routes/Laundry was transformed into server. I'm trying to modularize everything to make it easier to test. More refactoring to come...
