--- docker/docker-compose.yml
@@ -1,12 +1,0 @@
-version: '2'
-services:
-  laundry:
-    build: .
-    ports:
-     - "80:80"
-    depends_on:
-     - redis
-  redis:
-    image: redis
-    ports:
-    - "6379:6379"
