--- docker/laundry-api.conf
@@ -1,42 +1,0 @@
-# cups-server - docker container based local cups-server
-description "my docker cups-server"  
-author "Ian Holden"
-
-start on filesystem and started docker  
-stop on runlevel [!2345]
-
-# Automatically restart process if crashed (limit 3 times in 4 minutes)
-respawn limit 3 240
-
-# our process forks(clones) many times (maybe 10) so expect fork, or expect daemon are useless.
-# we use the technique explained here http://stackoverflow.com/questions/12200217/can-upstart-expect-respawn-be-used-on-processes-that-fork-more-than-twice
-
-# start the container in the pre-start script
-pre-start script  
-    # wait (if necessary) for our docker context to be accessible
-    while [ ! -f /home/ubuntu/purdue-laundry-api/docker-compose.yml ]
-    do
-      sleep 1
-    done
-    sleep 10 
-    /usr/bin/docker-compose -f /home/ubuntu/purdue-laundry-api/docker-compose.yml up -d
-end script
-
-# run a process that stays up while our docker container is up. Upstart will track this PID
-script  
-    sleepWhileAppIsUp(){
-        while docker ps | grep "$1" >/dev/null; do
-            sleep 2
-        done
-    }
-
-    sleepWhileAppIsUp "laundry"
-end script
-
-# stop docker container after the stop event has completed
-post-stop script  
-    if docker ps | grep laundry;
-    then
-        /usr/bin/docker-compose -f /home/ubuntu/purdue-laundry-api/docker-compose.yml stop
-    fi
-end script
