--- routes/Laundry/get_status.js
@@ -1,4 +1,0 @@
-module.exports = function (req, res) {
-	let status = ['Available', 'In Use', 'Almost Done', 'End of Cycle', 'Out of Order', 'Offline', 'Ready To Start'];
-	res.json(status);
-};
