+++ server/controllers/MachineController.js
@@ -1,5 +1,3 @@
-const parseHTML = require('./parse_html');
-const getLocations = require('./get_locations');
 const request = require('request');
 
 function getAllMachines(req) {
@@ -41,8 +39,7 @@ function getAllMachines(req) {
 	});
 }
 
-
-function getAllRoute(req, res) {
+function getMachines(req, res) {
 	console.time('allStart');
 	req.redis.exists('all', function (err, exists) {
 		if (exists === 0) {
@@ -62,4 +59,51 @@ function getAllRoute(req, res) {
 	});
 }
 
-module.exports = getAllRoute;
+function capitalizeFirstLetter(string) {
+	return string.charAt(0).toUpperCase() + string.slice(1);
+}
+
+function getMachinesAtLocation(req, res) {
+	req.logger.info({type: 'GET', location: req.params.location});
+	getURL(req.params.location, req).then(function (u) {
+		let url = capitalizeFirstLetter(u);
+		if (url === undefined) {
+			req.logger.err('Incorrect URL');
+			res.status(404).send('Error');
+			return;
+		}
+
+		req.redis.exists(req.params.location, function (err, exists) {
+			if (err) {
+				req.logger.err('Redis error- ' + err);
+			}
+
+			if (exists === 0) {
+				request(url, function (err, response, body) {
+					let results = [];
+					if (!err && response.statusCode === 200) {
+						results = parseHTML(body);
+					}
+					req.redis.set(req.params.location, JSON.stringify(results));
+					req.redis.expire(req.params.location, 60);
+					res.json(results);
+				});
+			} else {
+				req.redis.get(req.params.location, function (err, result) {
+					if (err) req.logger.err('Redis Error- ' + err);
+					let oldResponse = JSON.parse(result);
+					res.json(oldResponse);
+				});
+			}
+		});
+	});
+}
+
+function getPossibleStatuses(req, res) {
+	let status = ['Available', 'In Use', 'Almost Done', 'End of Cycle', 'Out of Order', 'Offline', 'Ready To Start'];
+	res.json(status);
+}
+
+module.exports = {
+	getMachines, getMachinesAtLocation, getPossibleStatuses
+};
