+++ server/lib/scraper.js
@@ -1,10 +1,10 @@
 const cheerio = require('cheerio');
 const request = require('request');
 
-module.exports = function getLocationURLs(req) {
+function scrapeLocations(redis) {
 	const url = 'http://wpvitassuds01.itap.purdue.edu/washalertweb/washalertweb.aspx';
 	return new Promise(function (resolve, reject) {
-		req.redis.get('location-urls', function (err, result) {
+		redis.get('location-urls', function (err, result) {
 			if (err) reject(err);
 			if (result) resolve(JSON.parse(result));
 			else {
@@ -19,11 +19,25 @@ module.exports = function getLocationURLs(req) {
 						};
 					});
 
-					req.redis.set('location-urls', JSON.stringify(locations));
-					req.redis.expire('location-urls', 1000 * 60 * 60 * 24);
+					redis.set('location-urls', JSON.stringify(locations));
+					redis.expire('location-urls', 1000 * 60 * 60 * 24);
 					resolve(locations);
 				});
 			}
 		});
 	});
-};
+}
+
+function getUrlFor(location, req) {
+	return scrapeLocations(req.redis).then(locations => locations.reduce((acc, item) => item.name === location ? item.url : ''));
+}
+
+function scrapeMachines(redis) {
+
+}
+
+function scrapeMachinesAt(location) {
+
+}
+
+module.exports = {scrapeLocations, getUrlFor, scrapeMachines, scrapeMachinesAt};
